Implement an `aggregate_utils.py` file to store shared aggregate processing functionality
